Moved "JwtUtils" of "authentication" project in an "utils" package
