+++ services/authentication/src/main/scala/it/cwmp/services/authentication/AuthenticationServiceVerticle.scala
@@ -5,6 +5,7 @@ import io.vertx.core.Handler
 import io.vertx.scala.ext.web.{Router, RoutingContext}
 import it.cwmp.services.VertxServer
 import it.cwmp.services.authentication.ServerParameters._
+import it.cwmp.services.authentication.utils.JwtUtils
 import it.cwmp.utils.Utils.{httpStatusNameToCode, stringToOption}
 import it.cwmp.utils.{HttpUtils, Logging}
